--- app/src/main/java/com/smashing/app/data/remote/dto/my/GetMyRecentReviewStatsResponse.kt
@@ -1,34 +1,0 @@
-package com.smashing.app.data.remote.dto.my
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class GetMyRecentReviewStatsResponse(
-    @SerialName("ratingCounts")
-    val ratingCounts: RatingCounts,
-    @SerialName("tagCounts")
-    val tagCounts: TagCounts,
-) {
-    @Serializable
-    data class RatingCounts(
-        @SerialName("best")
-        val best: Long,
-        @SerialName("good")
-        val good: Long,
-        @SerialName("bad")
-        val bad: Long,
-    )
-
-    @Serializable
-    data class TagCounts(
-        @SerialName("goodManner")
-        val goodManner: Long,
-        @SerialName("onTime")
-        val onTime: Long,
-        @SerialName("fairPlay")
-        val fairPlay: Long,
-        @SerialName("fastResponse")
-        val fastResponse: Long,
-    )
-}
