+++ app/src/main/java/com/smashing/app/data/remote/service/MyService.kt
@@ -2,7 +2,7 @@ package com.smashing.app.data.remote.service
 
 import com.smashing.app.data.remote.dto.BaseResponse
 import com.smashing.app.data.remote.dto.my.AddSportProfileRequest
-import com.smashing.app.data.remote.dto.my.GetMyRecentReviewStatsResponse
+import com.smashing.app.data.remote.dto.review.GetMyRecentReviewStatsResponse
 import com.smashing.app.data.remote.dto.my.GetMyTierProfileResponse
 import com.smashing.app.data.remote.dto.my.MyPageData
 import com.smashing.app.data.remote.dto.my.MyProfileSwitchRequest
@@ -27,8 +27,4 @@ interface MyService {
 
     @GET("/api/v1/users/me/profiles/tier")
     suspend fun getMyTierProfile(): BaseResponse<GetMyTierProfileResponse>
-
-
-    @GET("/api/v1/users/{userId}/reviews/summary")
-    suspend fun getMyRecentReviewStats(): BaseResponse<GetMyRecentReviewStatsResponse>
 }
