+++ app/src/main/java/com/smashing/app/data/remote/service/ReviewService.kt
@@ -2,6 +2,7 @@ package com.smashing.app.data.remote.service
 
 import com.smashing.app.data.remote.dto.BaseResponse
 import com.smashing.app.data.remote.dto.cursor.CursorDto
+import com.smashing.app.data.remote.dto.review.GetMyRecentReviewStatsResponse
 import com.smashing.app.data.remote.dto.review.GetUserRecentReviewListResponse
 import com.smashing.app.data.remote.dto.review.MyProfileReviewListData
 import retrofit2.http.GET
@@ -27,5 +28,9 @@ interface ReviewService {
         @Query("cursor") cursor: String?,
         @Query("size") size: Int?,
     ): BaseResponse<MyProfileReviewListData>
+
+    @GET("/api/v1/users/{userId}/reviews/summary")
+    suspend fun getMyRecentReviewStats(): BaseResponse<GetMyRecentReviewStatsResponse>
+
 
 }
